refactor(pipeline): extract config.py helpers to config_helpers (#521)
Move _derive_zone, _ssh_transport, _read_erp_user out of config.py (87
lines, over the 80-line pipeline cap) into a new sibling config_helpers.py.
config.py drops to 60 lines; build_config() unchanged. verify_cli.py's
import of _ssh_transport redirected to the new module — it was the only
external importer of a private helper. Pure refactor: helpers moved
byte-for-byte, no behavioral change.

fixes #521
