+++ tools/size_baselines.json
@@ -120,7 +120,8 @@
   "tools/pipeline/stages/__init__.py": 0,
   "tools/pipeline/stages/common/__init__.py": 0,
   "tools/pipeline/stages/common/ansible_output.py": 65,
-  "tools/pipeline/stages/common/config.py": 87,
+  "tools/pipeline/stages/common/config.py": 60,
+  "tools/pipeline/stages/common/config_helpers.py": 39,
   "tools/pipeline/stages/common/connectivity.py": 18,
   "tools/pipeline/stages/common/known_hosts.py": 69,
   "tools/pipeline/stages/common/log_format.py": 22,
